feat: migrating local database completely by SQLDelight

+++ feature/gratefulness/overview/src/commonMain/kotlin/com/joohnq/gratefulness/overview/presentation/GratefulnessOverviewViewModel.kt
@@ -57,7 +57,7 @@ class GratefulnessOverviewViewModel(
             }.launchIn(viewModelScope)
     }
 
-    private fun delete(id: Int) {
+    private fun delete(id: Long) {
         viewModelScope.launch {
             try {
                 deleteGratefulnessUseCase(id).getOrThrow()
