+++ feature/mood/overview/src/commonMain/kotlin/com/joohnq/overview/presentation/MoodOverviewViewModel.kt
@@ -45,7 +45,7 @@ class MoodOverviewViewModel(
         }
     }
 
-    private fun delete(id: Int) {
+    private fun delete(id: Long) {
         viewModelScope.launch {
             try {
                 deleteMoodUseCase(id).getOrThrow()
