+++ feature/sleep-quality/api/src/commonMain/kotlin/com/joohnq/sleep_quality/api/entity/DSleepQualityProperties.kt
@@ -3,6 +3,6 @@ package com.joohnq.sleep_quality.api.entity
 import com.joohnq.sleep_quality.api.property.SleepQualityProperties
 
 data class DSleepQualityProperties(
-    override val id: Int,
+    override val id: Long,
     override val level: Int,
 ) : SleepQualityProperties
